fix modinstall.py
